requirements.txt
alabaster==0.7.9
argh==0.26.2
Babel==2.3.4
backports-abc==0.5
certifi==2016.9.26
docutils==0.13.1
imagesize==0.7.1
Jinja2==2.8
livereload==2.5.0
MarkupSafe==0.23
pathtools==0.1.2
port-for==0.3.1
Pygments==2.1.3
pytz==2016.10
PyYAML==3.12
requests==2.12.3
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5
sphinx-autobuild==0.6.0
sphinx-rtd-theme==0.1.9
tornado==4.4.2
watchdog==0.8.3
pdfminer.six==20170720
